refactor: split model-collection.ts into modules
Extract 609-line monolithic file into 6 focused modules:
- types.ts: ModelGroup, FreeModel, PaidModel type definitions
- env.ts: environment-based model exclusion sets
- filters.ts: excludeModelsByName utility
- free-models.ts: RAW_FREE_MODELS data
- paid-models.ts: RAW_PAID_MODELS data
- index.ts: composed exports (FREE_MODELS, PAID_MODELS, MODEL_COLLECTION)

model-collection.ts is now a re-export barrel — 2 import sites unchanged.
